Create MyRunner.class
Add a command line runner to start taking user input and give it  a way to exit. This is a spring component which will automatically start when the spring application is started.
Change the test to include a mockbean to prevent the cli runner to start during tests and getting stuck in an infinite loop.